Limit inputs for rating

+++ src/main/java/seedu/address/model/deck/Difficulty.java
@@ -32,7 +32,7 @@ public class Difficulty {
     }
 
     public double getDifficulty() {
-        return numberOfAttempts == 0 ? 0 : totalRating / numberOfAttempts;
+        return numberOfAttempts == 0 ? 0 : (double) totalRating / (double) numberOfAttempts;
     }
 
     public int getTotalRating() {
